refactor: Change to relative imports to allow run from main
As run.py will be executed as the main program code was added to
facilitate this and the import were changed to relative paths to import
from the module directory.
