+++ run.py
@@ -1,3 +1,4 @@
-# Your code goes here.
-# You can delete these comments, but do not change the name of this file
-# Write your code to expect a terminal of 80 characters wide and 24 rows high
+from battleships.game import Game
+
+if __name__ == "__main__":
+    Game().new_game()
